requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.1
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.28.2
tqdm==4.65.0
urllib3==1.26.14
yarl==1.8.2
